% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_issues.R
\name{close_issue}
\alias{close_issue}
\title{Close an issue}
\usage{
close_issue(owner, repo, issue_number)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}

\item{issue_number}{Number id of the issue}
}
\description{
Close an issue
}
