% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_milestones.R
\name{create_milestone}
\alias{create_milestone}
\title{Create an issue on a repo from R list data}
\usage{
create_milestone(owner, repo, milestone_content)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}

\item{milestone_content}{R List with milestone content}
}
\description{
Create an issue on a repo from R list data
}
