% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_issues.R
\name{edit_issue}
\alias{edit_issue}
\title{Edit an issue}
\usage{
edit_issue(owner, repo, issue_number, issue_content)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}

\item{issue_number}{Number id of the issue}

\item{issue_content}{R List with issue content}
}
\description{
Edit an issue
}
