% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{find_tree_sha_in_commit}
\alias{find_tree_sha_in_commit}
\title{Title}
\usage{
find_tree_sha_in_commit(commit)
}
\arguments{
\item{commit}{Commit JSON already converted to R list}
}
\description{
Title
}
