% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_commit_by_sha}
\alias{get_commit_by_sha}
\title{Title}
\usage{
get_commit_by_sha(owner, repo, commit_sha)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}

\item{commit_sha}{Sha of a the commit of interest}
}
\description{
Title
}
