% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_files_from_repo}
\alias{get_files_from_repo}
\title{Get the names of all files in repo}
\usage{
get_files_from_repo(owner, repo)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}
}
\description{
Get the names of all files in repo
}
\examples{
get_files_from_repo("xvrdm", "ggrough")
}
