% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_files_from_tree}
\alias{get_files_from_tree}
\title{List all files included in a tree}
\usage{
get_files_from_tree(tree)
}
\arguments{
\item{tree}{Tree JSON already converted to R list}
}
\description{
This function will recursively also list the files in the trees
listed in the root tree and so on
}
