% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_files_from_tree_sha}
\alias{get_files_from_tree_sha}
\title{List all files included in a tree}
\usage{
get_files_from_tree_sha(tree_sha, src_dir = NULL, owner, repo)
}
\arguments{
\item{tree_sha}{Tree JSON already converted to R list}

\item{src_dir}{Directory path in which the tree files are}

\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}
}
\description{
This function will recursively also list the files in the trees
listed in the root tree and so on
}
