% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_last_commit}
\alias{get_last_commit}
\title{Get all data about the most recent commit}
\usage{
get_last_commit(owner, repo)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}
}
\description{
Get all data about the most recent commit
}
