% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_last_commits_summary}
\alias{get_last_commits_summary}
\title{Get a summary of the most recents commits}
\usage{
get_last_commits_summary(owner, repo)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}
}
\description{
Get a summary of the most recents commits
}
