% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_files.R
\name{get_tree_by_sha}
\alias{get_tree_by_sha}
\title{Title}
\usage{
get_tree_by_sha(owner, repo, tree_sha)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}

\item{tree_sha}{Sha of a the tree of interest}
}
\description{
Title
}
