% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{PATCH_gh}
\alias{PATCH_gh}
\title{Title}
\usage{
PATCH_gh(url, ..., res_type = "raw")
}
\arguments{
\item{url}{Relative endpoint url for Github API}

\item{...}{Named parameters to be sent as POST body content}

\item{res_type}{Type of data returned}
}
\description{
Title
}
