% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_get_issues.R
\name{create_issue}
\alias{create_issue}
\title{Title}
\usage{
create_issue(owner, repo, issue_content)
}
\arguments{
\item{owner}{Owner of the Github Repo (org or individual user)}

\item{repo}{Name of the Github Repo}

\item{issue_content}{R List with issue content}
}
\description{
Title
}
