% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{POST_gh}
\alias{POST_gh}
\title{Title}
\usage{
POST_gh(url, ...)
}
\arguments{
\item{url}{Relative endpoint url for Github API}

\item{...}{Named parameters to be sent as POST body content}
}
\description{
Title
}
